+++ src/clients/contracts/index.ts
@@ -1,11 +1,11 @@
 import { Abi, Account, Address, Chain, PublicClient, Transport, WalletClient } from 'viem';
 
-import { AavePositionMigrator } from '@/clients/contracts/AavePositionMigrator';
 import { Erc20 } from '@/clients/contracts/Erc20';
 import { OldERC721Wrapper } from '@/clients/contracts/OldERC721Wrapper';
+import { AavePositionMigrator } from '@/clients/contracts/PositionMigrator/AavePositionMigrator';
+import { UniswapV3PositionMigrator } from '@/clients/contracts/PositionMigrator/UniswapV3PositionMigrator';
 import { PurchaseBundlerV1 } from '@/clients/contracts/PurchaseBundlerV1';
 import { PurchaseBundlerV2 } from '@/clients/contracts/PurchaseBundlerV2';
-import { UniswapV3PositionMigrator } from '@/clients/contracts/UniswapV3PositionMigrator';
 import {
   getContracts,
   getVersionFromMslAddress,
